structure

+++ math/table.php
@@ -67,7 +67,9 @@ function print_header($text) {
 
 function print_text($text) {
 	echo "<div align='center'>\n";
+	echo "<h3>\n";
 	echo "$text\n";
+	echo "</h3>\n";
 	echo "</div>\n";
 }
 
@@ -131,13 +133,15 @@ function show_pdf_file($file) {
 }
 
 function show_link_file($header, $file, $button_text) {
-	echo "<div align='center'>\n";
-	echo "<span>\n";
-	echo "<a class='button' href='$file.tex' download> $header $button_text (.tex) </a>";
-	echo "</span>\n";
-	echo "<span>\n";
-	echo "<a class='button' href='$file.pdf' download> $header $button_text (.pdf) </a>";
-	echo "</span>\n";
-	echo "</div>\n";
+	if (file_exists($file.".pdf")) {
+		echo "<div align='center'>\n";
+		echo "<span>\n";
+		echo "<a class='button' href='$file.tex' download> $header $button_text (.tex) </a>";
+		echo "</span>\n";
+		echo "<span>\n";
+		echo "<a class='button' href='$file.pdf' download> $header $button_text (.pdf) </a>";
+		echo "</span>\n";
+		echo "</div>\n";
+	}
 }
 ?>
